+++ packages/mui-utils/src/useId.ts
@@ -1,11 +1,6 @@
 import * as React from 'react';
 
-/**
- * Prefer `useReactId` if the ID is only passed to React props
- * @param idOverride
- * @returns
- */
-export default function useId(idOverride?: string): string | undefined {
+function useRandomId(idOverride?: string): string | undefined {
   const [defaultId, setDefaultId] = React.useState(idOverride);
   const id = idOverride || defaultId;
   React.useEffect(() => {
@@ -17,4 +12,20 @@ export default function useId(idOverride?: string): string | undefined {
     }
   }, [defaultId]);
   return id;
+}
+
+/**
+ *
+ * @example <div id={useId} />
+ * @param idOverride
+ * @returns {string} Can only be passed to props
+ */
+export default function useReactId(idOverride?: string): string | undefined {
+  // TODO: Remove `React as any` once `useId` is part of stable types.
+  if ((React as any).useId !== undefined) {
+    const reactId = (React as any).useId();
+    return idOverride ?? reactId;
+  }
+  // eslint-disable-next-line react-hooks/rules-of-hooks -- `React.useId` is invariant at runtime.
+  return useRandomId(idOverride);
 }
